refactor: delete ProductService interface and adapt the controller
